factor out sanitization helpers
